requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyJWT==2.8.0
SQLAlchemy==1.4.56
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pdfplumber==0.8.0
python-docx==0.8.11
joblib==1.3.2
numpy==1.26.1
pandas==2.2.2
gspread==5.8.0
google-auth==2.23.0
slack-sdk==3.22.1
requests==2.31.0
